use revert_list_filters_to_tuple also in target. switch and upgrade tests
